fix: S29 round 3 code review polish — extract sceneColors, role cleanup
Addresses all LOW findings + one CONSIDER from claude-review bot round 3 on
PR #409. Bot verdict was already APPROVE — these are polish items, resolved
per CLAUDE.md rule 6b.

**LOW — Redundant onChange slice in SceneDetailSheet**
Both textareas already have `maxLength` enforced by the browser, so the
`slice(0, MAX_*)` inside onChange handlers never fires. Removed from both
direction and notes for consistency.

**LOW — Duplicate direction truncation across layers**
shotListFilters.groupShots stored `direction.slice(0, 100)` in ShotGroup, but
every presentation consumer (SceneHeader, SceneTableRow, SceneContextBanner)
re-truncated at 60 or 100 on render. Removed the data-layer slice — the raw
direction flows through and each renderer owns its own presentation.

**LOW — getSceneColor in component file creates component-to-component coupling**
Moved SCENE_COLORS, SceneColorKey, and getSceneColor from SceneHeader.tsx to
a new src-vnext/features/shots/lib/sceneColors.ts utility module. SceneHeader
re-exports for backward compat. Updated all 6 consumer imports
(shotColumnRenderers, SceneDetailSheet, RenumberShotsDialog, SceneTableRow,
SceneAssignPopover, SceneContextBanner).

**LOW — Redundant role="row" on tr elements**
`<tr>` inside `<tbody>` has an implicit ARIA role of "row". Removed explicit
role="row" from SceneTableRow and all 3 `<tr>` renderings in ShotsTable.

**CONSIDER — Scene number 0 accepted**
handleSceneNumberBlur previously accepted any non-negative integer including 0.
Scene 0 produces shot numbers like "0A", "0B" which sort before Scene 1 and
look odd. Tightened the validation to require >= 1 with an updated error
message, set min={1} on the Input, and added a Firestore rule guard:
`sceneNumber is int && sceneNumber >= 1`. createLane's auto-increment already
starts at 1 so this only affects manual override.

Tests: 167 files / 1978 passing. Lint zero. Build clean (12.17s).
