+++ firestore.rules
@@ -726,8 +726,9 @@ service cloud.firestore {
           (isAdmin() || isProducer());
       }
 
-      // Lanes (Scenes) — explicit rule with size validation on direction and notes.
-      // Must come BEFORE the wildcard catch-all so it takes precedence.
+      // Lanes (Scenes) — explicit rule with size validation on direction/notes
+      // and positive-integer validation on sceneNumber. Must come BEFORE the
+      // wildcard catch-all so it takes precedence.
       match /lanes/{laneId} {
         allow read: if clientMatches(clientId) &&
           (isAdmin() || producerCanAccessProject(clientId, projectId) ||
@@ -744,6 +745,11 @@ service cloud.firestore {
             !('notes' in request.resource.data) ||
             request.resource.data.notes == null ||
             (request.resource.data.notes is string && request.resource.data.notes.size() <= 5000)
+          ) &&
+          (
+            !('sceneNumber' in request.resource.data) ||
+            request.resource.data.sceneNumber == null ||
+            (request.resource.data.sceneNumber is int && request.resource.data.sceneNumber >= 1)
           );
         allow delete: if clientMatches(clientId) &&
           (isAdmin() || producerCanAccessProject(clientId, projectId) ||
