+++ src-vnext/features/shots/components/SceneContextBanner.tsx
@@ -1,4 +1,4 @@
-import { getSceneColor } from "@/features/shots/components/SceneHeader"
+import { getSceneColor } from "@/features/shots/lib/sceneColors"
 import type { Lane } from "@/shared/types"
 
 // ---------------------------------------------------------------------------
